changes according to comments
