+++ habitat_baselines/common/env_utils.py
@@ -7,29 +7,31 @@
 import random
 from typing import Type
 
-import numpy as np
-
 import habitat
 from habitat import Config, Env, VectorEnv, make_dataset
 
 
-def make_env_fn(config: Config, env_class: Type, rank: int) -> Env:
+def make_env_fn(
+    task_config: Config, rl_env_config: Config, env_class: Type, rank: int
+) -> Env:
     r"""
     Creates an env of type env_class with specified config and rank.
     This is to be passed in as an argument when creating VectorEnv.
     Args:
-        config: config file for creating env.
+        task_config: task config file for creating env.
+        rl_env_config: RL env config for creating env.
         env_class: class type of the env to be created.
-        rank: rank of env to be created.
+        rank: rank of env to be created (for seeding).
 
     Returns:
         env object created according to specification.
     """
-    dataset = make_dataset(config.DATASET.TYPE, config=config.DATASET)
-    config.defrost()
-    config.SIMULATOR.SCENE = dataset.episodes[0].scene_id
-    config.freeze()
-    env = env_class(config_env=config, config_baseline=config, dataset=dataset)
+    dataset = make_dataset(
+        task_config.DATASET.TYPE, config=task_config.DATASET
+    )
+    env = env_class(
+        config_env=task_config, config_baseline=rl_env_config, dataset=dataset
+    )
     env.seed(rank)
     return env
 
@@ -47,53 +49,54 @@ def construct_envs(config: Config, env_class: Type) -> VectorEnv:
     Returns:
         VectorEnv object created according to specification.
     """
-    baseline_cfg = config.TRAINER.RL.PPO
-    env_configs = []
-    env_classes = [env_class for _ in range(baseline_cfg.num_processes)]
+    trainer_config = config.TRAINER.RL.PPO
+    rl_env_config = config.TRAINER.RL
+    config = config.TASK_CONFIG  # exclude trainer-specific configs
+    env_configs, rl_env_configs = [], []
+    env_classes = [env_class for _ in range(trainer_config.num_processes)]
     dataset = make_dataset(config.DATASET.TYPE)
     scenes = dataset.get_scenes_to_load(config.DATASET)
 
     if len(scenes) > 0:
         random.shuffle(scenes)
 
-        assert len(scenes) >= baseline_cfg.num_processes, (
+        assert len(scenes) >= trainer_config.num_processes, (
             "reduce the number of processes as there "
             "aren't enough number of scenes"
         )
 
-    scene_splits = [[] for _ in range(baseline_cfg.num_processes)]
+    scene_splits = [[] for _ in range(trainer_config.num_processes)]
     for idx, scene in enumerate(scenes):
         scene_splits[idx % len(scene_splits)].append(scene)
 
     assert sum(map(len, scene_splits)) == len(scenes)
 
-    for i in range(baseline_cfg.num_processes):
+    for i in range(trainer_config.num_processes):
 
-        config_env = config.clone()
-        config_env.defrost()
+        env_config = config.clone()
+        env_config.defrost()
         if len(scenes) > 0:
-            config_env.DATASET.CONTENT_SCENES = scene_splits[i]
+            env_config.DATASET.CONTENT_SCENES = scene_splits[i]
 
-        config_env.SIMULATOR.HABITAT_SIM_V0.GPU_DEVICE_ID = (
-            baseline_cfg.sim_gpu_id
+        env_config.SIMULATOR.HABITAT_SIM_V0.GPU_DEVICE_ID = (
+            trainer_config.sim_gpu_id
         )
 
-        agent_sensors = baseline_cfg.sensors.strip().split(",")
-        for sensor in agent_sensors:
-            assert sensor in [
-                "RGB_SENSOR",
-                "DEPTH_SENSOR",
-            ], "currently 'RGB_SENSOR' and 'DEPTH_SENSOR' are supported "
-        config_env.SIMULATOR.AGENT_0.SENSORS = agent_sensors
-        config_env.freeze()
-        env_configs.append(config_env)
+        agent_sensors = trainer_config.sensors.strip().split(",")
+        env_config.SIMULATOR.AGENT_0.SENSORS = agent_sensors
+        env_config.freeze()
+        env_configs.append(env_config)
+        rl_env_configs.append(rl_env_config)
 
     envs = habitat.VectorEnv(
         make_env_fn=make_env_fn,
         env_fn_args=tuple(
             tuple(
                 zip(
-                    env_configs, env_classes, range(baseline_cfg.num_processes)
+                    env_configs,
+                    rl_env_configs,
+                    env_classes,
+                    range(trainer_config.num_processes),
                 )
             )
         ),
