+++ habitat_baselines/common/utils.py
@@ -14,7 +14,7 @@ import numpy as np
 import torch
 import torch.nn as nn
 
-from habitat import Config, get_config
+from habitat import Config
 from habitat.utils.visualizations.utils import images_to_video
 from habitat_baselines import BaseTrainer
 from habitat_baselines.common.baseline_registry import baseline_registry
@@ -22,7 +22,6 @@ from habitat_baselines.common.tensorboard_utils import (
     DummyWriter,
     TensorboardWriter,
 )
-from habitat_baselines.config.default import get_config as baseline_cfg
 
 # TODO distribute utilities in this file to separate files
 
@@ -40,26 +39,6 @@ def get_trainer(trainer_name: str, trainer_cfg: Config) -> BaseTrainer:
     trainer = baseline_registry.get_trainer(trainer_name)
     assert trainer is not None, f"{trainer_name} is not supported"
     return trainer(trainer_cfg)
-
-
-def get_exp_config(cfg_path: str, opts: List[str] = None) -> Config:
-    r"""
-    Create config object from path for a specific experiment run.
-    Args:
-        cfg_path: yaml config file path.
-        opts: list additional options or options to be overwritten.
-
-    Returns:
-        config object created.
-    """
-
-    config = Config(new_allowed=True)
-    config.merge_from_other_cfg(baseline_cfg(cfg_path))
-    print(config)
-    config.merge_from_other_cfg(get_config(config.TRAINER.RL.PPO.task_config))
-    if opts is not None:
-        config.merge_from_list(opts)
-    return config
 
 
 class Flatten(nn.Module):
@@ -105,29 +84,6 @@ def update_linear_schedule(optimizer, epoch, total_num_epochs, initial_lr):
     lr = initial_lr - (initial_lr * (epoch / float(total_num_epochs)))
     for param_group in optimizer.param_groups:
         param_group["lr"] = lr
-
-
-def experiment_args():
-    parser = argparse.ArgumentParser()
-    parser.add_argument(
-        "--run-type",
-        choices=["train", "eval"],
-        required=True,
-        help="run type of the experiment (train or eval)",
-    )
-    parser.add_argument(
-        "--exp-config",
-        type=str,
-        required=True,
-        help="path to config yaml containing info about experiment",
-    )
-    parser.add_argument(
-        "opts",
-        default=None,
-        nargs=argparse.REMAINDER,
-        help="Modify config options from command line",
-    )
-    return parser
 
 
 def batch_obs(observations: List[Dict]) -> Dict:
@@ -200,224 +156,11 @@ def generate_video(
     Returns:
         None
     """
-    print("======generating video======")
-    print(f"vidoe option: {config}")
     if config.video_option and len(images) > 0:
         video_name = f"episode{episode_id}_ckpt{checkpoint_idx}_spl{spl:.2f}"
         if "disk" in config.video_option:
             images_to_video(images, config.video_dir, video_name)
         if "tensorboard" in config.video_option:
-            print("writing to TB")
             tb_writer.add_video_from_np_images(
                 f"episode{episode_id}", checkpoint_idx, images, fps=fps
             )
-
-
-class RolloutStorage:
-    def __init__(
-        self,
-        num_steps,
-        num_envs,
-        observation_space,
-        action_space,
-        recurrent_hidden_state_size,
-    ):
-        self.observations = {}
-
-        for sensor in observation_space.spaces:
-            self.observations[sensor] = torch.zeros(
-                num_steps + 1,
-                num_envs,
-                *observation_space.spaces[sensor].shape,
-            )
-
-        self.recurrent_hidden_states = torch.zeros(
-            num_steps + 1, num_envs, recurrent_hidden_state_size
-        )
-
-        self.rewards = torch.zeros(num_steps, num_envs, 1)
-        self.value_preds = torch.zeros(num_steps + 1, num_envs, 1)
-        self.returns = torch.zeros(num_steps + 1, num_envs, 1)
-
-        self.action_log_probs = torch.zeros(num_steps, num_envs, 1)
-        if action_space.__class__.__name__ == "Discrete":
-            action_shape = 1
-        else:
-            action_shape = action_space.shape[0]
-
-        self.actions = torch.zeros(num_steps, num_envs, action_shape)
-        if action_space.__class__.__name__ == "Discrete":
-            self.actions = self.actions.long()
-
-        self.masks = torch.ones(num_steps + 1, num_envs, 1)
-
-        self.num_steps = num_steps
-        self.step = 0
-
-    def to(self, device):
-        for sensor in self.observations:
-            self.observations[sensor] = self.observations[sensor].to(device)
-
-        self.recurrent_hidden_states = self.recurrent_hidden_states.to(device)
-        self.rewards = self.rewards.to(device)
-        self.value_preds = self.value_preds.to(device)
-        self.returns = self.returns.to(device)
-        self.action_log_probs = self.action_log_probs.to(device)
-        self.actions = self.actions.to(device)
-        self.masks = self.masks.to(device)
-
-    def insert(
-        self,
-        observations,
-        recurrent_hidden_states,
-        actions,
-        action_log_probs,
-        value_preds,
-        rewards,
-        masks,
-    ):
-        for sensor in observations:
-            self.observations[sensor][self.step + 1].copy_(
-                observations[sensor]
-            )
-        self.recurrent_hidden_states[self.step + 1].copy_(
-            recurrent_hidden_states
-        )
-        self.actions[self.step].copy_(actions)
-        self.action_log_probs[self.step].copy_(action_log_probs)
-        self.value_preds[self.step].copy_(value_preds)
-        self.rewards[self.step].copy_(rewards)
-        self.masks[self.step + 1].copy_(masks)
-
-        self.step = (self.step + 1) % self.num_steps
-
-    def after_update(self):
-        for sensor in self.observations:
-            self.observations[sensor][0].copy_(self.observations[sensor][-1])
-
-        self.recurrent_hidden_states[0].copy_(self.recurrent_hidden_states[-1])
-        self.masks[0].copy_(self.masks[-1])
-
-    def compute_returns(self, next_value, use_gae, gamma, tau):
-        if use_gae:
-            self.value_preds[-1] = next_value
-            gae = 0
-            for step in reversed(range(self.rewards.size(0))):
-                delta = (
-                    self.rewards[step]
-                    + gamma * self.value_preds[step + 1] * self.masks[step + 1]
-                    - self.value_preds[step]
-                )
-                gae = delta + gamma * tau * self.masks[step + 1] * gae
-                self.returns[step] = gae + self.value_preds[step]
-        else:
-            self.returns[-1] = next_value
-            for step in reversed(range(self.rewards.size(0))):
-                self.returns[step] = (
-                    self.returns[step + 1] * gamma * self.masks[step + 1]
-                    + self.rewards[step]
-                )
-
-    def recurrent_generator(self, advantages, num_mini_batch):
-        num_processes = self.rewards.size(1)
-        assert num_processes >= num_mini_batch, (
-            "Trainer requires the number of processes ({}) "
-            "to be greater than or equal to the number of "
-            "trainer mini batches ({}).".format(num_processes, num_mini_batch)
-        )
-        num_envs_per_batch = num_processes // num_mini_batch
-        perm = torch.randperm(num_processes)
-        for start_ind in range(0, num_processes, num_envs_per_batch):
-            observations_batch = defaultdict(list)
-
-            recurrent_hidden_states_batch = []
-            actions_batch = []
-            value_preds_batch = []
-            return_batch = []
-            masks_batch = []
-            old_action_log_probs_batch = []
-            adv_targ = []
-
-            for offset in range(num_envs_per_batch):
-                ind = perm[start_ind + offset]
-
-                for sensor in self.observations:
-                    observations_batch[sensor].append(
-                        self.observations[sensor][:-1, ind]
-                    )
-
-                recurrent_hidden_states_batch.append(
-                    self.recurrent_hidden_states[0:1, ind]
-                )
-
-                actions_batch.append(self.actions[:, ind])
-                value_preds_batch.append(self.value_preds[:-1, ind])
-                return_batch.append(self.returns[:-1, ind])
-                masks_batch.append(self.masks[:-1, ind])
-                old_action_log_probs_batch.append(
-                    self.action_log_probs[:, ind]
-                )
-
-                adv_targ.append(advantages[:, ind])
-
-            T, N = self.num_steps, num_envs_per_batch
-
-            # These are all tensors of size (T, N, -1)
-            for sensor in observations_batch:
-                observations_batch[sensor] = torch.stack(
-                    observations_batch[sensor], 1
-                )
-
-            actions_batch = torch.stack(actions_batch, 1)
-            value_preds_batch = torch.stack(value_preds_batch, 1)
-            return_batch = torch.stack(return_batch, 1)
-            masks_batch = torch.stack(masks_batch, 1)
-            old_action_log_probs_batch = torch.stack(
-                old_action_log_probs_batch, 1
-            )
-            adv_targ = torch.stack(adv_targ, 1)
-
-            # States is just a (N, -1) tensor
-            recurrent_hidden_states_batch = torch.stack(
-                recurrent_hidden_states_batch, 1
-            ).view(N, -1)
-
-            # Flatten the (T, N, ...) tensors to (T * N, ...)
-            for sensor in observations_batch:
-                observations_batch[sensor] = self._flatten_helper(
-                    T, N, observations_batch[sensor]
-                )
-
-            actions_batch = self._flatten_helper(T, N, actions_batch)
-            value_preds_batch = self._flatten_helper(T, N, value_preds_batch)
-            return_batch = self._flatten_helper(T, N, return_batch)
-            masks_batch = self._flatten_helper(T, N, masks_batch)
-            old_action_log_probs_batch = self._flatten_helper(
-                T, N, old_action_log_probs_batch
-            )
-            adv_targ = self._flatten_helper(T, N, adv_targ)
-
-            yield (
-                observations_batch,
-                recurrent_hidden_states_batch,
-                actions_batch,
-                value_preds_batch,
-                return_batch,
-                masks_batch,
-                old_action_log_probs_batch,
-                adv_targ,
-            )
-
-    @staticmethod
-    def _flatten_helper(t: int, n: int, tensor: torch.Tensor) -> torch.Tensor:
-        r"""
-        Given a tensor of size (t, n, ..), flatten it to size (t*n, ...).
-        Args:
-            t: first dimension of tensor.
-            n: second dimension of tensor.
-            tensor: target tensor to be flattened.
-
-        Returns:
-            flattened tensor of size (t*n, ...)
-        """
-        return tensor.view(t * n, *tensor.size()[2:])
